Refactor: Carousel 컴포넌트 분리
