+++ src/components/index.js
@@ -1,4 +1,5 @@
-import Pagination from './Pagination'
-import Star from './Star'
-import Modal from './Modal'
-export { Pagination, Star, Modal }
+import Pagination from './Pagination';
+import Star from './Star';
+import Modal from './Modal';
+import Carousel from './Carousel';
+export { Pagination, Star, Modal, Carousel };
