+++ src/pages/Product/components/index.js
@@ -1,25 +1,23 @@
-import ProductSimpleView from './ProductSimpleView'
-import Option from './Option'
-import ProductNavigation from './ProductNavigation'
-import UserStylingView from './UserStylingView'
-import UserStylingPhoto from './UserStylingPhoto'
-import SelectedOption from './SelectedOption'
-import ProductInfo from './ProductInfo'
-import ProductSidebar from './ProductSidebar'
-import OptionSelectView from './OptionSelectView'
-import Review from './Review'
-import Ask from './Ask'
-import Delivery from './Delivery'
-import Recommend from './Recommend'
-import ReviewDetail from './ReviewDetail'
-import AskDetail from './AskDetail'
+import ProductSimpleView from './ProductSimpleView';
+import Option from './Option';
+import ProductNavigation from './ProductNavigation';
+import UserStylingView from './UserStylingView';
+import SelectedOption from './SelectedOption';
+import ProductInfo from './ProductInfo';
+import ProductSidebar from './ProductSidebar';
+import OptionSelectView from './OptionSelectView';
+import Review from './Review';
+import Ask from './Ask';
+import Delivery from './Delivery';
+import Recommend from './Recommend';
+import ReviewDetail from './ReviewDetail';
+import AskDetail from './AskDetail';
 
 export {
   ProductSimpleView,
   Option,
   ProductNavigation,
   UserStylingView,
-  UserStylingPhoto,
   SelectedOption,
   ProductInfo,
   ProductSidebar,
@@ -30,4 +28,4 @@ export {
   Recommend,
   ReviewDetail,
   AskDetail,
-}
+};
